delete with message by admin

+++ src/users/providers/users-del.provider.ts
@@ -4,6 +4,8 @@ import { User } from '../entities/user.entity';
 import { Repository } from 'typeorm';
 import * as bcrypt from 'bcryptjs';
 import { UsersGetProvider } from './users-get.provider';
+import { UserType } from '../../utils/enums';
+import { UsersOtpProvider } from './users-otp.provider';
 
 @Injectable()
 export class UsersDelProvider {
@@ -11,6 +13,7 @@ export class UsersDelProvider {
     @InjectRepository(User)
     private readonly usersRepository: Repository<User>,
     private readonly usersGetProvider: UsersGetProvider,
+    private readonly usersOtpProvider: UsersOtpProvider,
   ) {}
 
   async deleteMe(id: number, password: string) {
@@ -24,8 +27,12 @@ export class UsersDelProvider {
     return user;
   }
 
-  async deleteUserById(id: number) {
+  async deleteUserById(id: number, message: string) {
     const user = await this.usersGetProvider.findById(id);
+    if (user.userType === UserType.SUPER_ADMIN) {
+      throw new UnauthorizedException("You Can't");
+    }
+    await this.usersOtpProvider.sendSms(user.phone, message);
     await this.usersRepository.delete(id);
     return user;
   }
